Startin writing my own test server
The idea is to increase iteration speed by developing locally
and also to be able to run real unit tests.

So far, you can download a local copy of dynamodb and run that. make
sure the region is us-east-1. Then run node server.js in the test
directory. This should currently run one sample request/response.
It will create the table if this is the first time, and you'll see a
little `get error` message which is correct.