tests pass, working on how intents update database
So currently I don't have good control over what is saved to the
database. I have set methods that only set one attribute, which is
inefficient. I should be able to pass in any dict with whatever nested
attributes there may be, and have that written to the db.

The model I'm currently imagining is to pass around some
`peristant_attributes` dict. Initialize it with whatever is in
the database already. Then have intents fill that up and modify
that however they want. Then after the handlers are done save that to
the database. We can even be fancy and not do the set if the dict
hasn't been modified